Refine random_reference feature
