+++ snowfakery/data_generator_runtime.py
@@ -1,7 +1,6 @@
 from collections import defaultdict
 from datetime import date
 from contextlib import contextmanager
-from enum import Enum, auto
 
 from typing import Optional, Dict, List, Sequence, Mapping, NamedTuple
 
@@ -13,6 +12,7 @@ from .utils.template_utils import FakerTemplateLibrary
 from .template_funcs import StandardFuncs
 from .data_gen_exceptions import DataGenSyntaxError, DataGenNameError
 import snowfakery
+from snowfakery.object_rows import NicknameSlot, SlotState, ObjectRow
 
 OutputStream = "snowfakery.output_streams.OutputStream"
 ObjectTemplate = "snowfakery.data_generator_runtime_object_model.ObjectTemplate"
@@ -112,54 +112,6 @@ class Dependency(NamedTuple):
 yaml.SafeDumper.add_representer(
     Dependency, lambda representer, obj: representer.represent_list(obj)
 )
-
-
-class SlotState(Enum):
-    """The current state of a NicknameSlot.
-
-    UNUSED=empty, ALLOCATED=referenced, CONSUMED=object generated"""
-
-    UNUSED = auto()
-    ALLOCATED = auto()
-    CONSUMED = auto()
-
-
-class NicknameSlot:
-    """A slot that represents a Nickname or Tablename"""
-
-    _tablename: str
-    id_manager: IdManager
-    allocated_id: int = None
-
-    def __init__(self, tablename, id_manager):
-        self._tablename = tablename
-        self.id_manager = id_manager
-
-    @property
-    def id(self):
-        "Get an id corresponding to this slot. Generate one if necessary."
-        if self.allocated_id is None:
-            self.allocated_id = self.id_manager.generate_id(self._tablename)
-        return self.allocated_id
-
-    def consume_slot(self):
-        "Mark a slot as filled by an object and return its id for use by the object."
-        rc = self.allocated_id
-        self.allocated_id = SlotState.CONSUMED
-        return rc
-
-    @property
-    def status(self):
-        "Is the slot empty/unreferenced, allocated/referenced or consumed/used"
-        if self.allocated_id is None:
-            return SlotState.UNUSED
-        elif isinstance(self.allocated_id, int):
-            return SlotState.ALLOCATED
-        elif self.allocated_id == SlotState.CONSUMED:
-            return SlotState.CONSUMED
-
-    def __repr__(self):
-        return f"<NicknameSlot {vars(self)}>"
 
 
 class Globals(yaml.YAMLObject):
@@ -378,7 +330,7 @@ class RuntimeContext:
     easier. From the point of view of other modules this is "the interpreter"
     but internally its mostly just proxying to other classes."""
 
-    obj: Optional["ObjectRow"] = None
+    obj: Optional[ObjectRow] = None
     template_evaluator_recipe = JinjaTemplateEvaluatorFactory()
     current_template = None
 
@@ -529,46 +481,6 @@ def evaluate_function(func, args: Sequence, kwargs: Mapping, context):
         }
     value = func(*args, **kwargs)
     return value
-
-
-class ObjectRow(yaml.YAMLObject):
-    """Represents a single row
-
-    Uses __getattr__ so that the template evaluator can use dot-notation."""
-
-    yaml_loader = yaml.SafeLoader
-    yaml_dumper = yaml.SafeDumper
-    yaml_tag = "!snowfakery_objectrow"
-
-    __slots__ = ["_tablename", "_values", "_child_index"]
-
-    def __init__(self, tablename, values=(), index=0):
-        self._tablename = tablename
-        self._values = values
-        self._child_index = index
-
-    def __getattr__(self, name):
-        try:
-            return self._values[name]
-        except KeyError:
-            raise AttributeError(name)
-
-    def __str__(self):
-        return str(self.id)
-
-    def __repr__(self):
-        return f"<ObjectRow {self._tablename} {self.id}>"
-
-    def __getstate__(self):
-        return {"_tablename": self._tablename, "_values": self._values}
-
-    def __setstate__(self, state):
-        for slot, value in state.items():
-            setattr(self, slot, value)
-
-    @property
-    def _name(self):
-        return self._values.get("name")
 
 
 def output_batches(
