+++ snowfakery/output_streams.py
@@ -25,7 +25,7 @@ except (ImportError, ModuleNotFoundError) as e:
         raise exception
 
 
-from .data_generator_runtime import ObjectRow, NicknameSlot
+from .object_rows import ObjectRow, ObjectReference
 from .parse_recipe_yaml import TableInfo
 
 
@@ -55,7 +55,7 @@ class OutputStream(ABC):
         sourcetable: str,
         fieldname: str,
         source_row_dict: Dict,
-        target_object_row: Union[ObjectRow, NicknameSlot],
+        target_object_row: Union[ObjectRow, ObjectReference],
     ) -> Union[str, int]:
         return target_object_row.id
 
@@ -66,7 +66,7 @@ class OutputStream(ABC):
         pass
 
     def cleanup(self, field_name, field_value, sourcetable, row):
-        if isinstance(field_value, (ObjectRow, NicknameSlot)):
+        if isinstance(field_value, (ObjectRow, ObjectReference)):
             return self.flatten(sourcetable, field_name, row, field_value)
         else:
             encoder = self.encoders.get(type(field_value))
@@ -147,7 +147,7 @@ class DebugOutputStream(FileOutputStream):
         sourcetable: str,
         fieldname: str,
         source_row_dict: Dict,
-        target_object_row: Union[ObjectRow, NicknameSlot],
+        target_object_row: Union[ObjectRow, ObjectReference],
     ) -> Union[str, int]:
         return f"{target_object_row._tablename}({target_object_row.id})"
 
@@ -345,7 +345,7 @@ class GraphvizOutputStream(FileOutputStream):
         sourcetable: str,
         fieldname: str,
         source_row_dict: Dict,
-        target_object_row: Union[ObjectRow, NicknameSlot],
+        target_object_row: Union[ObjectRow, ObjectReference],
     ) -> Union[str, int]:
         source_node_name = self.generate_node_name(
             sourcetable, source_row_dict.get("name"), source_row_dict.get("id")
